Make this change optional
in the last commit this got commented out. that seems to cause issues on a few devices. so those few devices just set STOCK_ART := true in your BoardConfig.mk and you're set. I made it opt out instead of opt in because almost everybody has built with it the last change with no issues and it does seem to make a noticeable improvement and since its broken (no pun intended) far fewer devices then it has worked on it is opt out.